STT/requirements.txt
faster-whisper
numpy
sounddevice
scipy
webrtcvad-wheels
fastapi
uvicorn
python-multipart
requests
python-dotenv
